requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
alembic>=1.13.0
pydantic>=2.6.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
sqlalchemy-utils>=0.42.0
requests>=2.31.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
httpx>=0.24.0
pytest-cov>=3.0.0
coverage>=6.0.0
pytest-mock>=3.0.0
